evaluation/requirements.txt
torch
torchvision
pytorch-fid
numpy
pillow
